Pull generic crypto types into submodule #230
