+++ dams/Cargo.toml
@@ -41,3 +41,6 @@ webpki = "0.22"
 
 [build-dependencies]
 tonic-build = "0.8.0"
+
+[dev-dependencies]
+serde_json = "1"
